ci: use minor versions as upper bound in dependencies
the issue with #935 could happen because dependabot didn't check for minor version updates.

according to semantic versioning, deprecations can be introduced in minor version updates. there are basically two ways how we can treat deprecation warnings in CI:
- treat warnings as errors
- print warnings

We decided to treat warnings as errors some time ago and I still think that it's the most sensible approach, as I rarely read the warnings in CI tests. 

The problem is, that we specified major versions as upper bounds in `pyproject.toml`. dependabot would therefore only try to update the dependencies if a major version update happens. that's why the CI suddenly failed in unrelated new PRs, as a deprecation was introduced to polars.

In this PR I changed the upper bounds of the dependencies to minor versions. This means that we will get more dependabot spam again, but we're getting warned in time about deprecations in the dependabot PR, and unrelated PRs won't get any issues with deprecations.

If dependabot reveals some new deprecations, we can still decide in the dependabotPR if we want to apply the required changes in code or ignore the specific deprecation warning and create an issue.

Overall, I would say that it's really important that recent deprecations in dependencies do not mess with any development in unrelated PRs, so something like what happened to @saeub shouldn't happen again.

This PR should solve the issue.