ci: exclude all bots from contributor list
Apparently, `exclude-contributors` does not exclude the commits from the release draft. It only affects the contributor list at the end of the draft.

I would like to exclude all of our bots in the contributor list so I just added the other `dependabot` alias.

What I currently don't know how to achieve is completely excluding a commit from the changelog if they are of a particular author.